Bug fixed!
This was a tough one: when the OperationTable.data_init() was called, the
herp_oprec_t was not correctly initialized, so the OperationTable.fetch_*
commands were not working correctly.

To be precise, the OperationTable.fetch_external_id() is still not working
for external operations! This is not a problem for the moment, but the
OperationTable interface should totally be fixed!